Add Wanderer, Beast, Architect, Shaman crest splits.
Reaper had the transition split implemented in the continuous section -
changed to be a normal continuous split.

The witch crest appears to split when the cursed crest is obtained, so
the logic to determine if the "proper" witch crest is obtained will
require more work & be done in a follow-up.