Add Verdania Lake, Enter Verdania Castle, Enter Upper Wormways splits
I'm not entirely sure if "Verdania Castle" is the best term, but the
wiki uses it and I think calling it the Clover Dancers arena would be
misleading, as it doesn't involve the boss starting at all.

Upper Wormways is a low% request - it refers to the part that either
needs dash or a cocoon skip to pass through from the bottom.